factory-izing the quic server stack
