+++ source/common/http/http3/BUILD
@@ -19,6 +19,7 @@ envoy_cc_library(
         "//source/common/http:codec_client_lib",
         "//source/common/http:conn_pool_base_lib",
         "//source/common/quic:client_connection_factory_lib",
+        "@com_github_google_quiche//:quic_core_deterministic_connection_id_generator_lib",
     ],
 )
