+++ source/common/quic/codec_impl.h
@@ -5,8 +5,7 @@
 
 #include "source/common/common/assert.h"
 #include "source/common/common/logger.h"
-#include "source/common/quic/envoy_quic_client_session.h"
-#include "source/common/quic/envoy_quic_server_session.h"
+#include "source/common/quic/quic_filter_manager_connection_impl.h"
 
 namespace Envoy {
 namespace Quic {
@@ -29,56 +28,11 @@ public:
   Http::Protocol protocol() override { return Http::Protocol::Http3; }
   // Returns true if the session has data to send but queued in connection or
   // stream send buffer.
-  bool wantsToWrite() override;
+  bool wantsToWrite() override { return quic_session_.bytesToSend() > 0; }
 
 protected:
   QuicFilterManagerConnectionImpl& quic_session_;
   Http::Http3::CodecStats& stats_;
-};
-
-class QuicHttpServerConnectionImpl : public QuicHttpConnectionImplBase,
-                                     public Http::ServerConnection {
-public:
-  QuicHttpServerConnectionImpl(
-      EnvoyQuicServerSession& quic_session, Http::ServerConnectionCallbacks& callbacks,
-      Http::Http3::CodecStats& stats,
-      const envoy::config::core::v3::Http3ProtocolOptions& http3_options,
-      const uint32_t max_request_headers_kb, const uint32_t max_request_headers_count,
-      envoy::config::core::v3::HttpProtocolOptions::HeadersWithUnderscoresAction
-          headers_with_underscores_action);
-
-  // Http::Connection
-  void goAway() override;
-  void shutdownNotice() override;
-  void onUnderlyingConnectionAboveWriteBufferHighWatermark() override;
-  void onUnderlyingConnectionBelowWriteBufferLowWatermark() override;
-
-  EnvoyQuicServerSession& quicServerSession() { return quic_server_session_; }
-
-private:
-  EnvoyQuicServerSession& quic_server_session_;
-};
-
-class QuicHttpClientConnectionImpl : public QuicHttpConnectionImplBase,
-                                     public Http::ClientConnection {
-public:
-  QuicHttpClientConnectionImpl(EnvoyQuicClientSession& session,
-                               Http::ConnectionCallbacks& callbacks, Http::Http3::CodecStats& stats,
-                               const envoy::config::core::v3::Http3ProtocolOptions& http3_options,
-                               const uint32_t max_request_headers_kb,
-                               const uint32_t max_response_headers_count);
-
-  // Http::ClientConnection
-  Http::RequestEncoder& newStream(Http::ResponseDecoder& response_decoder) override;
-
-  // Http::Connection
-  void goAway() override;
-  void shutdownNotice() override {}
-  void onUnderlyingConnectionAboveWriteBufferHighWatermark() override;
-  void onUnderlyingConnectionBelowWriteBufferLowWatermark() override;
-
-private:
-  EnvoyQuicClientSession& quic_client_session_;
 };
 
 } // namespace Quic
