+++ source/extensions/filters/network/http_connection_manager/BUILD
@@ -2,7 +2,6 @@ load(
     "//bazel:envoy_build_system.bzl",
     "envoy_cc_extension",
     "envoy_extension_package",
-    "envoy_select_enable_http3",
 )
 
 licenses(["notice"])  # Apache 2
@@ -38,36 +37,34 @@ envoy_cc_extension(
         "//source/common/filter:config_discovery_lib",
         "//source/common/http:conn_manager_lib",
         "//source/common/http:default_server_string_lib",
+        "//source/common/http:filter_chain_helper_lib",
         "//source/common/http:request_id_extension_lib",
         "//source/common/http:utility_lib",
-        "//source/common/http:filter_chain_helper_lib",
         "//source/common/http/http1:codec_lib",
         "//source/common/http/http1:settings_lib",
         "//source/common/http/http2:codec_lib",
         "//source/common/http/http3:codec_stats_lib",
         "//source/common/json:json_loader_lib",
         "//source/common/local_reply:local_reply_lib",
+        "//source/common/network:cidr_range_lib",
         "//source/common/protobuf:utility_lib",
         "//source/common/router:rds_lib",
         "//source/common/router:scoped_rds_lib",
         "//source/common/runtime:runtime_lib",
+        "//source/common/tracing:custom_tag_lib",
         "//source/common/tracing:http_tracer_lib",
         "//source/common/tracing:http_tracer_manager_lib",
         "//source/common/tracing:tracer_config_lib",
-        "//source/common/tracing:custom_tag_lib",
-        "//source/common/network:cidr_range_lib",
         "//source/extensions/filters/network:well_known_names",
         "//source/extensions/filters/network/common:factory_base_lib",
-        "//source/extensions/http/original_ip_detection/xff:config",
         "//source/extensions/http/header_validators/envoy_default:config",
+        "//source/extensions/http/original_ip_detection/xff:config",
         "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/filters/network/http_connection_manager/v3:pkg_cc_proto",
-        "@envoy_api//envoy/extensions/request_id/uuid/v3:pkg_cc_proto",
-        "@envoy_api//envoy/extensions/http/original_ip_detection/xff/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/http/header_validators/envoy_default/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/http/original_ip_detection/xff/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/request_id/uuid/v3:pkg_cc_proto",
         "@envoy_api//envoy/type/tracing/v3:pkg_cc_proto",
         "@envoy_api//envoy/type/v3:pkg_cc_proto",
-    ] + envoy_select_enable_http3([
-        "//source/common/quic:codec_lib",
-    ]),
+    ],
 )
