+++ test/common/quic/BUILD
@@ -159,11 +159,11 @@ envoy_cc_test(
         ":test_proof_source_lib",
         ":test_utils_lib",
         "//envoy/stats:stats_macros",
-        "//source/common/quic:codec_lib",
         "//source/common/quic:envoy_quic_alarm_factory_lib",
         "//source/common/quic:envoy_quic_connection_helper_lib",
         "//source/common/quic:envoy_quic_server_connection_lib",
         "//source/common/quic:envoy_quic_server_session_lib",
+        "//source/common/quic:server_codec_lib",
         "//source/server:configuration_lib",
         "//test/mocks/event:event_mocks",
         "//test/mocks/http:http_mocks",
@@ -186,7 +186,7 @@ envoy_cc_test(
     deps = [
         ":test_utils_lib",
         "//envoy/stats:stats_macros",
-        "//source/common/quic:codec_lib",
+        "//source/common/quic:client_codec_lib",
         "//source/common/quic:envoy_quic_alarm_factory_lib",
         "//source/common/quic:envoy_quic_client_connection_lib",
         "//source/common/quic:envoy_quic_client_session_lib",
