+++ test/integration/BUILD
@@ -882,7 +882,8 @@ envoy_cc_test_library(
         "@envoy_api//envoy/config/listener/v3:pkg_cc_proto",
     ] + envoy_select_enable_http3([
         "//source/common/quic:active_quic_listener_lib",
-        "//source/common/quic:quic_factory_lib",
+        "//source/common/quic:quic_client_factory_lib",
+        "//source/common/quic:quic_server_factory_lib",
         "@com_github_google_quiche//:quic_test_tools_session_peer_lib",
     ]),
 )
@@ -1931,7 +1932,8 @@ envoy_cc_test(
             ":protocol_integration_test_lib",
             "//source/common/quic:active_quic_listener_lib",
             "//source/common/quic:client_connection_factory_lib",
-            "//source/common/quic:quic_factory_lib",
+            "//source/common/quic:quic_client_factory_lib",
+            "//source/common/quic:quic_server_factory_lib",
             "//source/common/quic:quic_transport_socket_factory_lib",
             "//test/integration/filters:pause_filter_for_quic_lib",
         ],
@@ -1971,7 +1973,8 @@ envoy_cc_test(
         "//source/common/quic:envoy_quic_client_session_lib",
         "//source/common/quic:envoy_quic_connection_helper_lib",
         "//source/common/quic:envoy_quic_proof_verifier_lib",
-        "//source/common/quic:quic_factory_lib",
+        "//source/common/quic:quic_client_factory_lib",
+        "//source/common/quic:quic_server_factory_lib",
         "//test/common/quic:test_utils_lib",
         "//test/integration/filters:encoder_decoder_buffer_filter_lib",
         "//test/integration/filters:stream_info_to_headers_filter_lib",
