+++ test/integration/fake_upstream.cc
@@ -17,7 +17,7 @@
 #include "source/common/network/utility.h"
 
 #ifdef ENVOY_ENABLE_QUIC
-#include "source/common/quic/codec_impl.h"
+#include "source/common/quic/server_codec_impl.h"
 #include "quiche/quic/test_tools/quic_session_peer.h"
 #endif
